SAK-21553 - Fixing user email search
Note that this does not allow the addition of external guest reviewers, it
just allows searching by registered users by email address. The existing
workflow is by design. If it should be changed (to work similarly as
viewing a presentation, where guests are allowed to be created on the fly),
a new ticket should be opened with a specific institutional need presented.