SAK-22335
Tidy in body-only mode eats <script> tags as they are moved the the non-existent <head>.

It is still useful for developers to run their portlet output through Tidy to clean it up.

So the code is not wholly removed - it is advisory only with the following properties changes.

portal.portlet.tidy.warnings=false 
Still false by default - when true - it runs Tidy in noisy mode. It does *not* take the 
output of Tidy - just checks the portlet output and then uses it unchanged. 

portal.portlet.tidy=false 
This is now ignored. We cannot use the output of Tidy as the markup. 

Does this seem illogical to you? Yes. Tell it to the Tidy folks...