SAK-11485 - Fix course grades with grouped assignments
This is part two of the fix. It wires up the beans so the external assessments
can be checked and checks visibility when computing the total points available
for a given student.

Note that there were some strange Spring-Hibernate transaction errors when
attempting to use the GradebookExternalAssessmentService directly, so there is
a new proxy bean for it that is marked to be non-transactional and read-only.
This is coarser-grained than I would like, but it is appropriate enough for the
need here.