SAK-22974 - Use LIKE instead of equals for Oracle support.
The CITATION_COLLECTION.PROPERTY_VALUE column is a CLOB on Oracle and
LONGTEXT on MySQL. SAK-22296 added a comparison of this column with a
literal using the equals operator, which is not supported on Oracle.
However, a LIKE works on both, and there are already two columns
specified in the WHERE clause, so the PROPERTY_VALUE set is tiny.