SAK-23733 - Fix external grade integration to show unprovided tool items
A bug was introduced that would not allow externally maintained gradebook
items to be shown unless the tool implemented a provider. The default
behavior should be that a provider need only be implemented if the tool
wishes to exert some additional controls, such as group access, on which
items are included for specific student views and grades.

This change implements a solution that does this by asking providers to
also implement a method to list all items they are responsible for. This
list is compared to the items that should be visible and the items that
are not reported by any provider to preserve the default behavior that
was present in the 2.8 release.