BLTI-230
Tighten up a few things.   Get the callback at the end of  registration to work.
Get the state updated.   A few other tweaks.   This basically completes the 
LTI 2.0 install workflow.  Many cleanup issues still to do but Sakai has the 
distinction of being the first LMS on the planet with an LTI 2.x implementation
checked into a source repo - be it ever so humble :)