BLTI-264
Add a second check for scheme mis-match in the generated HTML that auto-submits
the form.   This (a) checks if we are already the top windows and (b) if we are
in a frame compared the action of the form and the page/s href - the page is https
and the action is http, slip in a target="_blank" into the form.

This should work for LessonBuilder and any other situation where launch is 
in an iframe without any checking.  This is a nice safety net.