DASH-277
Attempt to add click handler to each list item.  However, handler is unable to trigger the click on the item's link without causing an endless loop.  In effect, the trigger registers as a click on the item that contains the link.  Find some way to get the link's action and add that to the list item's click handler.