STAT-360 Update easymock to 3.2.
In easymock 3.2 mocks are threadsafe by default so we shouldn't get tests failing intermitently.
Easymock 3+ also check that you are mocking sensible things so the expect calls on non-mock objects which were wrong have been removed.
The classextension dependency is no longer needed as well as this is now part of core easymock.