SAK-26139 Don't attempt to lookup aliases for null.
The timeout handler in the portal doesn't have a active site and so asks the portal to render a response with a null site, the portal tries to lookup aliases for null which will never work. The SiteService has a null check shortcut but the alias service doesn't. Adding shortcut to prevent nulls getting to the alias service.