SAK-28162
Refactor the portal code to place the portal's good copy of jQuery
1.9.1 in the variable $PBJQ (Peanut Butter and JQuery) and modify
portal scripts to access via $PDJQ instead of jQuery and $ 

We leave jQuery and $ as whatever the tool prefers so if they 
define global functions and then use them - their assumptions
hold.

One small fix that got fixed whilst testing the namespace things 
was to copy the portal.chat.video setup code from the neo portal to 
the morpheus portal.