SAK-25773
Restore the portal copy of jQuery namespacing code.  This looks like a
lot because the files get copied lots of places.  Also this is post a 
run of the SCSS compile step so a lot of files got changed as well.

The actual changes really happenned in the folders:

  library/src/webapp/skin/morpheus-master/js/src/*
  library/src/webapp/skin/morpheus-master/js/plugins/*

The rest of the changes are "knock on" effect of auto-generated files.